# projects/datasite/cmd/datasite/openapi.yaml
openapi: 3.0.3
info:
  title: Datasite API
  version: "1.0.0"
servers:
  - url: /api/v1
security:
  - admin_api_key: []
paths:
  /cinemas:
    get:
      operationId: listCinemas
      summary: List all cinemas
      responses:
        "200":
          $ref: "#/components/responses/CinemaList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createCinema
      summary: Create a cinema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CinemaInput"
      responses:
        "201":
          $ref: "#/components/responses/Cinema"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
  /cinemas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: getCinema
      summary: Get a cinema by ID
      responses:
        "200":
          $ref: "#/components/responses/Cinema"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateCinema
      summary: Replace a cinema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CinemaInput"
      responses:
        "200":
          $ref: "#/components/responses/Cinema"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: patchCinema
      summary: Partially update a cinema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CinemaPatch"
      responses:
        "200":
          $ref: "#/components/responses/Cinema"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteCinema
      summary: Delete a cinema
      responses:
        "204":
          description: Cinema deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
  /movie_logs:
    get:
      operationId: listMovieLogs
      summary: List all movie logs
      responses:
        "200":
          $ref: "#/components/responses/MovieLogList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createMovieLog
      summary: Create a movie log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieLogInput"
      responses:
        "201":
          $ref: "#/components/responses/MovieLog"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
  /movie_logs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: getMovieLog
      summary: Get a movie log by ID
      responses:
        "200":
          $ref: "#/components/responses/MovieLog"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateMovieLog
      summary: Replace a movie log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieLogInput"
      responses:
        "200":
          $ref: "#/components/responses/MovieLog"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: patchMovieLog
      summary: Partially update a movie log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieLogPatch"
      responses:
        "200":
          $ref: "#/components/responses/MovieLog"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteMovieLog
      summary: Delete a movie log
      responses:
        "204":
          description: Movie log deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        default:
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    admin_api_key:
      type: http
      scheme: bearer
      description: Admin API key passed as a Bearer token.
  schemas:
    Cinema:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    CinemaInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    CinemaPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
    MovieLog:
      type: object
      required: [id, movie_id, review_contains_spoilers]
      properties:
        id:
          type: integer
          format: int64
        movie_id:
          type: integer
          format: int64
        date:
          type: integer
          format: int64
          nullable: true
        cinema_id:
          type: integer
          format: int64
          nullable: true
        rating:
          type: number
          format: double
          nullable: true
        review:
          type: string
          nullable: true
        review_contains_spoilers:
          type: boolean
    MovieLogInput:
      type: object
      required: [movie_id]
      properties:
        movie_id:
          type: integer
          format: int64
        date:
          type: integer
          format: int64
          nullable: true
        cinema_id:
          type: integer
          format: int64
          nullable: true
        rating:
          type: number
          format: double
          nullable: true
        review:
          type: string
          nullable: true
        review_contains_spoilers:
          type: boolean
    MovieLogPatch:
      type: object
      properties:
        movie_id:
          type: integer
          format: int64
        date:
          type: integer
          format: int64
          nullable: true
        cinema_id:
          type: integer
          format: int64
          nullable: true
        rating:
          type: number
          format: double
          nullable: true
        review:
          type: string
          nullable: true
        review_contains_spoilers:
          type: boolean
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
  responses:
    Cinema:
      description: Cinema response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Cinema"
    CinemaList:
      description: List of cinemas
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Cinema"
    MovieLog:
      description: Movie log response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MovieLog"
    MovieLogList:
      description: List of movie logs
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/MovieLog"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Request conflicts with current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...